Enemies major update 1
Extended the IEnemy Interface.
Created an enemies manager wich control all the enemies of the scene.
With the EnemiesManager is possible to: alert other enemies of the same area, disable enemies when they die without destroy them, respown all the enemy of the area when pass throw the EnemiesAreaTrigger collider.